Extracted common code between aws_byte_cursor_read_be* proof harnesses
* Added aws_byte_cursor_read_common.h to proof_helpers directory, containing common code between all harnesses.
* Modified individual harnesses to call aws_byte_cursor_read_common with the appropriate definitions.
